requirements.txt
gdown==5.2.0
matplotlib==3.9.0  # Versi yang lebih kompatibel
numpy==1.21.4     # Versi numpy yang lebih stabil
onnxruntime==1.20.1
pandas==2.2.3
Pillow==11.2.1
pyiqa==0.1.13
streamlit==1.45.0
streamlit_image_comparison==0.0.4
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
wandb==0.19.10
